create test_utils module in core to share test-only helpers
