+++ core/src/lib.rs
@@ -5,3 +5,6 @@ pub mod report;
 pub mod parsers;
 
 pub mod error;
+
+#[cfg(any(test, feature = "testing"))]
+mod test_utils;
